pyproject.toml
[project]
name = "bdr"
version = "0.1.0"
description = "Business Directory REST API"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1,<6.0",
    "djangorestframework>=3.15",
    "djangorestframework-simplejwt>=5.3",
    "django-unfold>=0.40",
    "django-redis>=5.4",
    "psycopg2-binary>=2.9",
    "python-slugify>=8.0",
    "openpyxl>=3.1",
    "python-decouple>=3.8",
    "django-cors-headers>=4.4",
]

[project.optional-dependencies]
dev = [
    "ipython",
]

[tool.uv]
dev-dependencies = [
    "ipython",
]

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
